Implemented Tensor Normalization Funcs + Restructed file system for Tensors Ops
Motivation: Aside from new features needing to be added, there was an issue with the way Tensor operations were structed within all_element_ops.go and axis_ops.go. Before this commit, operations that worked in place on a Tensor and those that collapse an axis were all in the same source file for axis and all ops respectively. This was becoming cluttered so collapsing functions and in place functions into their own files for axis and all ops alike. These changes were reflected in the test files as well.